% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_functions.R
\name{create_analysis_data_frame}
\alias{create_analysis_data_frame}
\title{Munge and save RedCap data}
\usage{
create_analysis_data_frame(input_data, settings)
}
\arguments{
\item{settings}{}
}
\description{
Munge and save RedCap data
}
