% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieval_functions.R
\name{retrieve_redcap_data}
\alias{retrieve_redcap_data}
\title{Retrieve and save RedCap data}
\usage{
retrieve_redcap_data()
}
\arguments{
\item{token}{RedCap token}
}
\description{
Retrieve and save RedCap data
}
