% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_check_to_file}
\alias{write_check_to_file}
\title{Get a tasp dataset}
\usage{
write_check_to_file(intro_text, records = NULL, action_text = NULL,
  retrieval_date)
}
\arguments{
\item{intro_text}{some text}

\item{records}{some data}

\item{action_text}{some text}

\item{retrieval_date}{self-explanatory}
}
\description{
Get a tasp dataset
}
